+++ aws-java-sdk-iotanalytics/src/main/java/com/amazonaws/services/iotanalytics/model/DescribeChannelResult.java
@@ -29,6 +29,12 @@ public class DescribeChannelResult extends com.amazonaws.AmazonWebServiceResult<
      * </p>
      */
     private Channel channel;
+    /**
+     * <p>
+     * Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     * </p>
+     */
+    private ChannelStatistics statistics;
 
     /**
      * <p>
@@ -71,6 +77,47 @@ public class DescribeChannelResult extends com.amazonaws.AmazonWebServiceResult<
     }
 
     /**
+     * <p>
+     * Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     * </p>
+     * 
+     * @param statistics
+     *        Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     */
+
+    public void setStatistics(ChannelStatistics statistics) {
+        this.statistics = statistics;
+    }
+
+    /**
+     * <p>
+     * Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     * </p>
+     * 
+     * @return Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the
+     *         request.
+     */
+
+    public ChannelStatistics getStatistics() {
+        return this.statistics;
+    }
+
+    /**
+     * <p>
+     * Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     * </p>
+     * 
+     * @param statistics
+     *        Statistics about the channel. Included if the 'includeStatistics' parameter is set to true in the request.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public DescribeChannelResult withStatistics(ChannelStatistics statistics) {
+        setStatistics(statistics);
+        return this;
+    }
+
+    /**
      * Returns a string representation of this object; useful for testing and debugging.
      *
      * @return A string representation of this object.
@@ -82,7 +129,9 @@ public class DescribeChannelResult extends com.amazonaws.AmazonWebServiceResult<
         StringBuilder sb = new StringBuilder();
         sb.append("{");
         if (getChannel() != null)
-            sb.append("Channel: ").append(getChannel());
+            sb.append("Channel: ").append(getChannel()).append(",");
+        if (getStatistics() != null)
+            sb.append("Statistics: ").append(getStatistics());
         sb.append("}");
         return sb.toString();
     }
@@ -101,6 +150,10 @@ public class DescribeChannelResult extends com.amazonaws.AmazonWebServiceResult<
             return false;
         if (other.getChannel() != null && other.getChannel().equals(this.getChannel()) == false)
             return false;
+        if (other.getStatistics() == null ^ this.getStatistics() == null)
+            return false;
+        if (other.getStatistics() != null && other.getStatistics().equals(this.getStatistics()) == false)
+            return false;
         return true;
     }
 
@@ -110,6 +163,7 @@ public class DescribeChannelResult extends com.amazonaws.AmazonWebServiceResult<
         int hashCode = 1;
 
         hashCode = prime * hashCode + ((getChannel() == null) ? 0 : getChannel().hashCode());
+        hashCode = prime * hashCode + ((getStatistics() == null) ? 0 : getStatistics().hashCode());
         return hashCode;
     }