+++ aws-java-sdk-stepfunctions/src/main/java/com/amazonaws/services/stepfunctions/AWSStepFunctionsClientBuilder.java
@@ -27,7 +27,7 @@ import com.amazonaws.client.AwsSyncClientParams;
 @Generated("com.amazonaws:aws-java-sdk-code-generator")
 public final class AWSStepFunctionsClientBuilder extends AwsSyncClientBuilder<AWSStepFunctionsClientBuilder, AWSStepFunctions> {
 
-    private static final ClientConfigurationFactory CLIENT_CONFIG_FACTORY = new ClientConfigurationFactory();
+    private static final ClientConfigurationFactory CLIENT_CONFIG_FACTORY = new com.amazonaws.services.stepfunctions.AWSStepFunctionsClientConfigurationFactory();
 
     /**
      * @return Create new instance of builder with all defaults set.